Fix JNI type mismatch for Hysteria.start
Gomobile binds the Go `int` type to Java `long` to ensure compatibility across architectures.
This commit explicitly casts the Android File Descriptor (Int) to Long in `DTechVpnService.kt` to resolve the `Type mismatch: inferred type is Int but Long was expected` compilation error.

This is the final fix required for the build to pass.

+++ app/src/main/java/com/dtech/vpn/DTechVpnService.kt
@@ -96,7 +96,8 @@ class DTechVpnService : VpnService() {
                 // FIX: Run in background thread to prevent crash
                 Thread {
                     try {
-                        Hysteria.start(vpnInterface!!.fd, serverStr, auth, "")
+                        // Fix: Cast Int to Long for gomobile compatibility
+                        Hysteria.start(vpnInterface!!.fd.toLong(), serverStr, auth, "")
                         broadcastLog("Hysteria Core Connected Successfully!")
                     } catch (e: Exception) {
                         Log.e("DTechVPN", "Native Error", e)
